backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas==2.1.3
geopandas==0.14.1
shapely==2.0.2
requests==2.31.0
python-dotenv==1.0.0
groq==0.4.1
google-generativeai==0.3.2
numpy==1.26.2
pyproj==3.6.1
rtree==1.1.0

